sleep thread for refresh_rate duration if amortization is finished and let offset be of type long

+++ server_date/src/server_date.h
@@ -35,7 +35,7 @@ private:
     int sample_count;
 
     double precision;
-    double offset;
+    long offset;
 
     int counter_for_refresh = 0;
     int refresh_rate;
